update: Integration of common layouts
