+++ frontend/src/app/App.tsx
@@ -1,17 +1,20 @@
-import { Route, Routes } from "react-router-dom";
+import { Route, Routes, Navigate } from "react-router-dom";
 import LoginForm from "../pages/LoginForm";
 import Register from "../pages/register";
 import Dashboard from "../pages/Dashboard";
 
 const App = () => {
-  return(
+  return (
     <Routes>
-      <Route path="/" element = {<LoginForm/>}/>
-      <Route path="/dashboard" element = {<Dashboard/>}/>
-      <Route path="/register" element = {<Register/>}/>
-      <Route path="/dashboard" element = {<Dashboard/>}/>
+      {/* ログインページ */}
+      <Route path="/" element={<LoginForm />} />
+      <Route path="/register" element={<Register />} />
+      <Route path="/dashboard" element={<Dashboard />} />
+
+      {/* その他の未定義ルートは "/" にリダイレクト */}
+      <Route path="*" element={<Navigate to="/" replace />} />
     </Routes>
-  )
-}
+  );
+};
 
 export default App;
